Removes old Parquet support
